replace regex parsing with nom
I'm not sure if I actually like this better because it gives pretty
mediocre error messages, but I really wanted to try playing with making
a useful parser with nom and tonight was the night to try that out.

The good news is it works and passes all my existing tests, and
I actually understand all of the code here.